requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
supabase>=2.0.0
razorpay>=1.4.0
